pyproject.toml
[project]
name = "acropolis-bot"
version = "1.0.0"
description = "Advanced Polymarket trading bot with micro-arbitrage and multi-market support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.42.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "websockets>=12.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "aiofiles>=23.2.1",
    "web3>=6.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
]
